Make denoise optional, default to off
Optionally using bregman denoising on masks can give real significant
compression improvements, at the speed of computational costs.

Currently we are strugging on the computational side, so let's default
to off.

For the record, for the first 15 pages of
https://archive.org/details/100factsonsharks0000park the compression
ratio without denoise is 6.86; the compression ratio with denoise is
14.94.

The same is true for the first 10 pages of
https://archive.org/details/lanjingdeyanjing0000bing the compression
ratio without denoise is 14.83; the compression ratio with denoise is
22.57;