Make recode a library call
This should make it easier to pass argument metadata like '--', since
argparse has no sensible way to pass that as value (which makes sense,
to be honest).

Later on I will do some more reworking of internetarchivepdf/recode.py
to make it more library-esque and less like a call to a
executable/binary (clean up).