pdfrenderer: deal with different ppi and fontsizes in a better way
There are various scenarios that we have to deal with:

1. Tesseract 5 generated hOCR with the correct DPI embedded in the hOCR
   file.
2. Tesseract 5 generated hOCR with the incorrect (default) DPI embedded
   in the hOCR file (no DPI known at OCR time, for example), font sizes
   here are usually too large.
3. Tesseract 4 generated hOCR with no DPI information, font sizes here
   are also often too large.
4. hOCR files generated from Abbyy files using abbyy-to-hocr from
   archive-hocr-tools
5. Custom uploaded hOCR files

In the case of (1), we have to increase the font size for our PDFs.
In the case of (2, 3), we only have to adjust the font size for the
document size in the PDF.
In the case of (4), depending the version we used to generate the hOCR
file, we might not have any font information at all, in which case the
default font size can be a poor pick, so we estimate the font size from
the line bounding box height - the same applies to (5), in case those
files also do not have font sizes.