mrc: use dpi for sauvola, rework hocr mask
The binarisation results from Sauvola are better if we adjust the window
size to the DPI, so let's do that if we know the DPI.

Also rework the hOCR mask generation. The old code was quite slow and
relied on a lot of noise estimation for dealing with inverted text, but
didn't give the required results - this new implementation works better,
provided the OCR picked up the inverted text. This also provides a
~20-30% speed boost by making the hocr_mask_gen function many times
faster.