pdf-metadata-json: add ppi guards
In some cases the PDFs contain high resolution images that are rendered
as a very tiny image with respect to the page size, which could cause
our PPI estimator to return unwieldy numbers.

Additionally, if the image transform's .a and .d are both negative, this
means the image is rotated and we need to swap the width and height -
this also fixes the ppi calculation in some cases.

Finally, for good measure, let's put an upper limit of 1200 ppi on the
estimated ppi value.