pdf-metadata-json: deal with odd page rects
Some PDFs have pages with insane page rectangles, which in turn would
create a crazily inflated estimated_default_render_res. Detect this and
adjust the scale appropriatell to get to a sensible
estimated_default_render_res.

And in case some PDF pages are actually really large, we still want to
limit them. In this case, the max image size is 25000 x 25000.